refactor: ArticlePreviewItem 타입 파일 구조 정리(#170)
