+++ src/app/(main)/(home)/types/byIdeologySection.ts
@@ -1,4 +1,4 @@
-import { type ArticleBaseTypes } from '@/shared/types/articleItemType';
+import { type ArticleBaseTypes } from '@/shared/components/articlePreviewItem/types';
 
 export type IdeologyType = 'progressive' | 'center' | 'conservative';
