+++ src/app/(main)/(home)/types/hotArticleSection.ts
@@ -1,4 +1,4 @@
-import { type ArticleBaseTypes } from '@/shared/types/articleItemType';
+import { type ArticleBaseTypes } from '@/shared/components/articlePreviewItem/types';
 
 export type IdeologyType = 'progressive' | 'conservative';
