+++ src/app/(main)/search/types/search.ts
@@ -1,4 +1,4 @@
-import { SearchArticlePreviewItemTypes } from '@/shared/types/articleItemType';
+import { SearchArticlePreviewItemTypes } from '@/shared/components/articlePreviewItem/types';
 
 // search 결과 list 중 NewsResultList에 해당하는 타입
 export type NewsArticleItemData = {
